backend/requirements.txt
psycopg2-binary
markdown
Django==4.2.11
gunicorn==21.2.0
psycopg2==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mongoengine==0.28.2
drf-nested-routers==0.93.5
django-filter==24.1.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-debug-toolbar==4.3.0
pillow==10.2.0
factory-boy==3.2.1
Faker==15.3.4
pytest-django==4.5.2
model_bakery==1.12.0